Change cast in MediatorTests.cpp, Mediator.h, and Mediator.inl

+++ speed-dreams/source-2.2.3/src/simulatedDrivingAssistance/backend/Mediator.h
@@ -53,7 +53,7 @@ public:
         return &m_decisionMaker;
     }
 
-    void SetTicks(int tick)
+    void SetTicks(unsigned long tick)
     {
         m_tickCount = tick;
     }
